sql: destroy_tenant queues gc job
Previously, destory_tenant marked the tenant as inactive. A seperate
call to gc_tenant was required to trigger the deletion. Now, the
gc is scheduled by the destroy_tenant command. The gc_tenant command
will be deleted once this change is live in the Cockroach Cloud
Serverless clusters and all inactive tenants are deleted.

Release note: None